refactor(session): split lockworker to greeterworker lockworker and shutdownworker
